+++ examples/Serialization/validate_config.py
@@ -70,7 +70,9 @@ CONFIG = str(
         },
         "recommender": {
             "type": "SequentialGreedyRecommender",
-            "surrogate_model_cls": "GP",
+            "surrogate_model": {
+                "type": "GaussianProcessSurrogate"
+            },
             "acquisition_function_cls": "qEI"
         },
         "allow_repeated_recommendations": false,
@@ -133,7 +135,9 @@ INVALID_CONFIG = str(
         },
         "recommender": {
             "type": "SequentialGreedyRecommender",
-            "surrogate_model_cls": "GP",
+            "surrogate_model": {
+                "type": "GaussianProcessSurrogate"
+            },
             "acquisition_function_cls": "qEI"
         },
         "allow_repeated_recommendations": false,